SR 195 Imrpove ServerSideRedirect security
git-svn-id: svn://svn.code.sf.net/p/isfdb/code-svn/trunk@696 c56fabd3-1acd-401a-a842-f99f0872de9c

+++ mod/resolve_bad_url.py
@@ -10,16 +10,21 @@
 #     Date: $Date$
 
 from isfdb import *
-from common import *
-from isfdblib import *
 from SQLparsing import *
-from library import *
+from library import ServerSideRedirect
+from login import User
 
 	
 if __name__ == '__main__':
 
         pub_id = SESSION.Parameter(0, 'int')
 
+	user = User()
+	user.load()
+	user.load_moderator_flag()
+	if not user.moderator:
+                SESSION.DisplayError('Only Moderators Can Resolve Bad URLs')
+
         update = 'delete from bad_images where pub_id=%d' % pub_id
 	db.query(update)
-	ServerSideRedirect("http:/%s/mod/bad_images.cgi" % HTFAKE)
+	ServerSideRedirect('http:/%s/mod/bad_images.cgi' % HTFAKE)
